+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to `Tools for Apache Kafka®` are documented in this file.
 - Validation for available topics in `.kafka` files. See [#153](https://github.com/jlandersen/vscode-kafka/issues/153).
 - Simplify snippets. See [#180](https://github.com/jlandersen/vscode-kafka/pull/180).
 - Hover support in `.kafka` files. See [#149](https://github.com/jlandersen/vscode-kafka/issues/149).
+- String encoding serialization support. See [#181](https://github.com/jlandersen/vscode-kafka/issues/181).
 
 ## [0.12.0] - 2021-04-26
 ### Added